Rewrote the ELF generator to use the new functions
The ELF generator now composes a simple ARMv7 executable, composed by
three different simple code frames, with the main one calling the two
others.

Something like this [main]->[write]->[main]->[exit]

The symbols are not kept in the ELF executable though.

Still, the executable runs fine on an ARMv7 machine so...

It's time to build a nice little interface for our PoC Graphical
Assembler and see how it fares.

Signed-off-by: Myy <myy@miouyouyou.fr>